+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsApi.kt
@@ -98,8 +98,8 @@ internal class AnalyticsApi(
             .putOpt(FPTI_KEY_CONTEXT_TYPE, if (event.isVaultRequest) "BA-TOKEN" else "EC-TOKEN")
             .putOpt(FPTI_KEY_PAYPAL_ATTEMPTED_APP_SWITCH, event.didSdkAttemptAppSwitch)
             .putOpt(FPTI_KEY_FUNDING_SOURCE, event.fundingSource)
-            .putOpt(FPTI_KEY_IS_PURCHASE, event.isPurchase)
-            .putOpt(FPTI_KEY_IS_BILLING_AGREEMENT, event.isBillingAgreement)
+            .putOpt(FPTI_KEY_IS_PURCHASE, event.isPurchaseFlow)
+            .putOpt(FPTI_KEY_IS_BILLING_AGREEMENT, event.shouldRequestBillingAgreement)
             .putOpt(FPTI_KEY_BILLING_PLAN_TYPE, event.recurringBillingPlanType)
     }