+++ BraintreeCore/src/test/java/com/braintreepayments/api/core/AnalyticsApiUnitTest.kt
@@ -53,8 +53,8 @@ class AnalyticsApiUnitTest {
         contextId = "context-id",
         linkType = "link-type",
         isVaultRequest = true,
-        isPurchase = true,
-        isBillingAgreement = true,
+        isPurchaseFlow = true,
+        shouldRequestBillingAgreement = true,
         recurringBillingPlanType = "billing_plan_type",
         startTime = 123,
         endTime = 234,
@@ -78,8 +78,8 @@ class AnalyticsApiUnitTest {
         contextId = "context-id",
         linkType = "link-type",
         isVaultRequest = false,
-        isPurchase = true,
-        isBillingAgreement = true,
+        isPurchaseFlow = true,
+        shouldRequestBillingAgreement = true,
         recurringBillingPlanType = "billing_plan_type",
         startTime = 123,
         endTime = 234,
@@ -232,8 +232,8 @@ class AnalyticsApiUnitTest {
                                     "button_position": "${tokenizationKeyEvent.buttonOrder}",
                                     "context_type": "EC-TOKEN",
                                     "paypal_app_switch_url_received": ${tokenizationKeyEvent.didPayPalServerAttemptAppSwitch},
-                                    "is_purchase": ${tokenizationKeyEvent.isPurchase},
-                                    "is_billing_agreement": ${tokenizationKeyEvent.isBillingAgreement},
+                                    "is_purchase": ${tokenizationKeyEvent.isPurchaseFlow},
+                                    "is_billing_agreement": ${tokenizationKeyEvent.shouldRequestBillingAgreement},
                                     "billing_plan_type": "${tokenizationKeyEvent.recurringBillingPlanType}"
                                 }
                             ]
@@ -294,8 +294,8 @@ class AnalyticsApiUnitTest {
                                     "button_position": "${clientTokenEvent.buttonOrder}",
                                     "context_type": "BA-TOKEN",
                                     "paypal_app_switch_url_received": ${clientTokenEvent.didPayPalServerAttemptAppSwitch},
-                                    "is_purchase": ${tokenizationKeyEvent.isPurchase},
-                                    "is_billing_agreement": ${tokenizationKeyEvent.isBillingAgreement},
+                                    "is_purchase": ${tokenizationKeyEvent.isPurchaseFlow},
+                                    "is_billing_agreement": ${tokenizationKeyEvent.shouldRequestBillingAgreement},
                                     "billing_plan_type": "${tokenizationKeyEvent.recurringBillingPlanType}"
                                 }
                             ]