+++ BraintreeCore/src/test/java/com/braintreepayments/api/core/AnalyticsParamRepositoryUnitTest.kt
@@ -28,8 +28,8 @@ class AnalyticsParamRepositoryUnitTest {
         sut.didSdkAttemptAppSwitch = true
         sut.didEnablePayPalAppSwitch = true
         sut.fundingSource = expectedFundingSource
-        sut.isPurchase = true
-        sut.isBillingAgreement = true
+        sut.isPurchaseFlow = true
+        sut.shouldRequestBillingAgreement = true
         sut.recurringBillingPlanType = expectedBillingPlanType
     }
 
@@ -51,8 +51,8 @@ class AnalyticsParamRepositoryUnitTest {
         assertEquals(true, sut.didEnablePayPalAppSwitch)
         assertEquals(true, sut.didSdkAttemptAppSwitch)
         assertEquals(expectedFundingSource, sut.fundingSource)
-        assertEquals(true, sut.isPurchase)
-        assertEquals(true, sut.isBillingAgreement)
+        assertEquals(true, sut.isPurchaseFlow)
+        assertEquals(true, sut.shouldRequestBillingAgreement)
         assertEquals(expectedBillingPlanType, sut.recurringBillingPlanType)
 
         sut.reset()
@@ -62,8 +62,8 @@ class AnalyticsParamRepositoryUnitTest {
         assertNull(sut.didEnablePayPalAppSwitch)
         assertNull(sut.didSdkAttemptAppSwitch)
         assertNull(sut.fundingSource)
-        assertNull(sut.isPurchase)
-        assertNull(sut.isBillingAgreement)
+        assertNull(sut.isPurchaseFlow)
+        assertNull(sut.shouldRequestBillingAgreement)
         assertNull(sut.recurringBillingPlanType)
     }
 }