+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalPaymentAuthRequestParams.kt
@@ -30,12 +30,12 @@ data class PayPalPaymentAuthRequestParams @JvmOverloads internal constructor(
         get() = payPalRequest.getFundingSource().value
 
     @get:RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    val isPurchase: Boolean
+    val isPurchaseFlow: Boolean
         get() = payPalRequest is PayPalCheckoutRequest &&
                 payPalRequest.amount.toDoubleOrNull()?.let { it > 0 } == true
 
     @get:RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    val isBillingAgreement: Boolean
+    val shouldRequestBillingAgreement: Boolean
         get() = payPalRequest is PayPalCheckoutRequest && payPalRequest.shouldRequestBillingAgreement
 
     @get:RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)