+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree Android SDK Release Notes
 
+## unreleased
+
+* PayPal
+    * Pass `is_billing_agreement`, `is_purchase`, and `billing_plan_type` to analytics events.
+
 ## 5.23.0 (2026-02-12)
 
 * Venmo
@@ -7,8 +12,6 @@
 * DataCollector
     * Bump Magnes version to `5.6.0`
     * Add `collectDeviceDataOnSuccess` method that ensures Magnes SDK completes data submission prior to returning device data
-* PayPal
-    * Pass `is_billing_agreement`, `is_purchase`, and `billing_plan_type` to analytics events.
 
 ## 5.22.0 (2026-01-29)